[♻️refactor]: post api 함수 분리 리팩토링
