+++ src/pages/Plant/PlantDiary/PlantDiary.tsx
@@ -1,17 +1,15 @@
-import { useMutation } from '@tanstack/react-query';
-import { AxiosResponse } from 'axios';
 import { useEffect, useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { useNavigate, useParams } from 'react-router';
-import { toast } from 'react-toastify';
 
 import BreadcrumbAndTitle from '@/components/common/Breadcrumb/BreadcrumbAndTitle';
 import Spinner from '@/components/common/Spinner/Spinner';
-import { usePlantById, usePlantDiary } from '@/hooks/api/usePlant';
-import { API_AUTHORITY, API_ENDPOINT } from '@/services/apiEndpoint';
-import { axiosInstance } from '@/services/axiosInstance';
+import {
+  usePlantById,
+  usePlantDiary,
+  usePostPlantDiary,
+} from '@/hooks/api/usePlant';
 import { formatPlantDiaryDate } from '@/utils/formatDate';
-import { getUserToken } from '@/utils/getUserToken';
 
 import PlantCalendar from './components/PlantCalendar';
 import PlantDiaryForm from './PlantDiaryForm/PlantDiaryForm';
@@ -20,16 +18,25 @@ export default function PlantDiary() {
   const { id } = useParams();
   const navigate = useNavigate();
   const [selectDate, setSelectDate] = useState<Date>(new Date());
-  //? DiaryForm State
   const [content, setContent] = useState<string>('');
 
-  //? React Hook Form
+  //? API
+  const { data: plantData, isLoading: isPlantLoading } = usePlantById(
+    Number(id)
+  );
+  const { data: plantDiaryData, isLoading: isDiaryLoading } = usePlantDiary(
+    Number(id),
+    formatPlantDiaryDate(selectDate)
+  );
+  const { mutate: postPlantData } = usePostPlantDiary();
+
   const { register, handleSubmit, setValue } = useForm({
     defaultValues: {
       growth: '',
     },
   });
 
+  //? React Hook Form
   const onSubmit = handleSubmit(data => {
     postPlantData({
       id: Number(id),
@@ -38,48 +45,6 @@ export default function PlantDiary() {
       record: formatPlantDiaryDate(selectDate),
     });
   });
-
-  //? -------------- API 요청 함수 ---------------
-  const { data: plantData, isLoading: isPlantLoading } = usePlantById(
-    Number(id)
-  );
-  const { data: plantDiaryData, isLoading: isDiaryLoading } = usePlantDiary(
-    Number(id),
-    formatPlantDiaryDate(selectDate)
-  );
-
-  const postPlantDiary = async ({ id, growth, content, record }) => {
-    const token = getUserToken();
-    const body = { growth, content, record };
-
-    const res: AxiosResponse<string> = await axiosInstance.post(
-      `${API_AUTHORITY.PLANT}${API_ENDPOINT.PLANT.POST_PLANT_DIARY}/${id}`,
-      body,
-      {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      }
-    );
-
-    return res?.data;
-  };
-
-  const usePostPlantDiary = () => {
-    return useMutation({
-      mutationFn: postPlantDiary,
-      onSuccess: (message: string) => {
-        toast.success(message);
-      },
-      onError: () => {
-        toast.error('저장에 실패했습니다. 다시 시도해주세요.');
-      },
-    });
-  };
-
-  const { mutate: postPlantData } = usePostPlantDiary();
-
-  //? ------------------------------------------
 
   useEffect(() => {
     setContent(plantDiaryData?.content ?? '');
